Create GetVMEnvironmentTagsADOTask.yml
This is a small powershell task I put together to run in Azure DevOps when I have multiple VMs inside an Environment and I want to use tags to find an attribute from a specific VM in the environment. Think 3 tier app, where you want to pass the name of one VM for use in configuring an application during deployment. Handling in this way means the same set of deployment steps can be used across multiple environments within the same pipeline e.g. DEV/UAT/PROD